Made the InputManager script execute before the normal scripts
I got an null reference error pointing to the InputManager when running the game. The InputManager creates an InputActionMap in the Awake function that other scripts rely upon. The problem was that the other scripts OnEnable/OnDisable was executing before the InputManagers Awake function, which meant that they didn't get the references they needed.