Added a scene switching script and a level manager script to allow the player to move from one scene to another.
When the player walks into a trigger with the SceneSwitcher script, the LevelManager disables player input and starts loading the new scene. It also shows a VERY simple loading screen. The level manager doesn't allow the new scene to spawn before the loading screen has finished its transition. When the loading screen is done, the LevelManager allows the new scene to spawn. When the new scene spawns the loading screen hides itself and player input is enabled.