Made the dialogue system not need a singleton GameObject to work..
Moved most of the code from the DialogueManager into the DialogueBox script. Now the DialogueManager only acts as a message bus for the dialogue system. The reason I did this was to avoid having the dialogue sytem break if somebody forgot to a GameObject with the DialogueManager to the scene.