Setup 2 simple scenes to test the quest system
The system works but I don't have a way of saving game state between scenes. If I pick up a object in 1 scene and move to another scene and back the object will reappear. Currently I'm using the QuestStateChangedObserver to disable objects that are "picked up" by disabling them if the player enters a new scene and have completed the required quest.